project_1_cdc_lakehouse_and_dbt_analytics_pipeline/streaming/spark/requirements.txt
# Requirements for Spark Package
PyYAML>=6.0
pytest>=7.0
